requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.0
httpx>=0.27
openai>=1.30
pyyaml>=6.0
python-multipart>=0.0.9
